requirements.txt
pygame
python-dotenv
screeninfo
pysqlite3
typing-extensions
json